Merged Adapter with Activity
Renamed the repo to stopwatch-android-java-eclipse and the project to
stopwatch-android-java.

Renamed the controller package to android.

Moved the activity back into the android package, merged it with the
controller, and removed the StopwatchController interface.

Used a final class with a private constructor, like in
http://www.javapractices.com/topic/TopicAction.do?Id=2., for constants.

Renamed the state methods to onStartStop etc. to indicate that they are
application-level semantic events.

Updated states. Removed ResetState and added LapStopped state.